format variable that parsing color parameters

+++ OceanTokens/Classes/OceanExtensions/Shadows/OceanExtensionShadow.swift
@@ -26,7 +26,10 @@ extension UIView {
   public func applyShadow(parameters: Ocean.shadow.ShadowParameters){
       self.layer.shadowOffset = CGSize(width: parameters["x"] as! Int, height: parameters["y"] as! Int)
       self.layer.shadowRadius = CGFloat(parameters["radius"] as! Int)
-      let color = UIColor(red: (parameters["red"] as! CGFloat / 255), green: (parameters["green"] as! CGFloat  / 255), blue: (parameters["blue"] as! CGFloat / 255), alpha: parameters["alpha"] as! CGFloat)
+      let color = UIColor(red: CGFloat(parameters["red"] as! Int) / 255,
+                        green: CGFloat(parameters["green"] as! Int)  / 255,
+                        blue: CGFloat(parameters["blue"] as! Int) / 255,
+                        alpha: CGFloat(parameters["alpha"] as! Double))
       self.layer.shadowColor = color.cgColor
       self.layer.shadowOpacity = 1
   }
